+++ autoblocks/_impl/datasets/client.py
@@ -10,10 +10,14 @@ from typing import Optional
 
 import httpx
 
+from autoblocks._impl.api.exceptions import APIError
+from autoblocks._impl.api.exceptions import ValidationError
+from autoblocks._impl.api.exceptions import parse_error_response
+from autoblocks._impl.api.utils.helpers import build_path
+from autoblocks._impl.api.utils.serialization import deserialize_model
+from autoblocks._impl.api.utils.serialization import deserialize_model_list
+from autoblocks._impl.api.utils.serialization import serialize_model
 from autoblocks._impl.config.constants import API_ENDPOINT_V2
-from autoblocks._impl.datasets.exceptions import APIError
-from autoblocks._impl.datasets.exceptions import ValidationError
-from autoblocks._impl.datasets.exceptions import parse_error_response
 from autoblocks._impl.datasets.models.dataset import Dataset
 from autoblocks._impl.datasets.models.dataset import DatasetItem
 from autoblocks._impl.datasets.models.dataset import DatasetItemsSuccessResponse
@@ -21,10 +25,6 @@ from autoblocks._impl.datasets.models.dataset import DatasetListItem
 from autoblocks._impl.datasets.models.dataset import DatasetSchema
 from autoblocks._impl.datasets.models.dataset import SuccessResponse
 from autoblocks._impl.datasets.models.schema import create_schema_property
-from autoblocks._impl.datasets.utils.helpers import build_path
-from autoblocks._impl.datasets.utils.serialization import deserialize_model
-from autoblocks._impl.datasets.utils.serialization import deserialize_model_list
-from autoblocks._impl.datasets.utils.serialization import serialize_model
 from autoblocks._impl.util import cuid_generator
 
 log = logging.getLogger(__name__)
